added new statsfragment for on trip

+++ TestApplication/app/src/main/java/com/example/epic/testapplication/AsyncResponse.java
@@ -4,6 +4,6 @@ package com.example.epic.testapplication;
  * Created by henryshangguan on 6/25/15.
  */
 public interface AsyncResponse {
-    void processFinish(Integer[] output);
+    void processFinish(double[] output);
 }
